Implement separate mouse state for LiteLoader that preserves existing behavior
